\docType{package}
\name{boralis-package}
\alias{boralis}
\title{boralis: Extract useful objects from multivariate analyses in boral}
\description{Extract results from boral analyses and plot using ggplot2}
\section{Functions}{
	\strong{Data extraction}
  \itemize{
		\item \code{\link{boral_coefs}} Extract coefficients from a boral model
  }
	\strong{Plotting}
	\itemize{
		\item \code{\link{boral_coefsplot}} Draw a coefficient plot using ggplot2
	}
}